Implemented database methods for locations

+++ app/src/main/java/de/umpocketmind/FunctionalityClasses/Location.java
@@ -5,13 +5,13 @@ package de.umpocketmind.FunctionalityClasses;
  */
 public class Location {
 
-    private int id;
+    private long id;
     private String name;
     private String description;
     private double longtitude;
     private double latitude;
 
-    public Location(int id, String name, String description, double longtitude, double latitude) {
+    public Location(long id, String name, String description, double longtitude, double latitude) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -19,10 +19,10 @@ public class Location {
         this.latitude = latitude;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
     public String getName() {
